Completing of writing rest-full web service

+++ src/main/java/main/entities/User.java
@@ -9,28 +9,23 @@ import java.util.Date;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private long id;
     @Column(unique=true)
     private String email;
-
     @Column(name="first_name")
     private String firstName;
-
     @Column(name="last_name")
     private String lastName;
-
     @Column(name = "birth_date", columnDefinition = "DATE")
     private LocalDate birthDate;
     private String address;
-
     @Column(name = "phone",unique = true)
     private String phoneNumber;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
